pyproject.toml
[project]
name = "bachelors-recommendation-system"
version = "0.1.0"
description = "Movie recommendation system using collaborative filtering, content-based filtering, and two-tower models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "polars>=0.19.0",
  "pyarrow>=14.0.0",
  "numpy>=1.24,<2.0",
  "pandas>=2.1.0",
  "scipy>=1.11.0",
  "scikit-learn>=1.3.0",
  "scikit-surprise>=1.1.4",
  "sentence-transformers>=2.2.0",
  "torch>=2.0.0",
  "transformers>=4.35.0",
  "jupyter>=1.0.0",
  "ipykernel>=6.25.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "joblib>=1.3.0",
  "kagglehub>=0.2.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
    "pre-commit>=3.0.0",
    "pip-audit>=2.0.0",
]

[tool.black]
line-length = 127
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | data
)/
'''

[tool.isort]
profile = "black"
line_length = 127
skip_gitignore = true
